+++ src/main/java/roomescape/global/auth/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.global.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
@@ -13,7 +13,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.auth.domain.dto.TokenInfoDto;
+import roomescape.global.auth.domain.dto.TokenInfoDto;
 import roomescape.global.exception.AuthorizationException;
 
 @Component
