+++ src/main/java/roomescape/global/auth/domain/dto/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package roomescape.auth.domain.dto;
+package roomescape.global.auth.domain.dto;
 
 public record TokenResponseDto(String accessToken) {
 }
