+++ src/main/java/roomescape/global/config/WebMvcConfiguration.java
@@ -1,13 +1,13 @@
-package roomescape.config;
+package roomescape.global.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.JwtTokenProvider;
-import roomescape.config.interceptor.CheckAdminInterceptor;
-import roomescape.config.resolver.LoginMemberArgumentResolver;
+import roomescape.global.auth.JwtTokenProvider;
+import roomescape.global.config.interceptor.CheckAdminInterceptor;
+import roomescape.global.config.resolver.LoginMemberArgumentResolver;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
