+++ src/main/java/roomescape/global/config/interceptor/CheckAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.config.interceptor;
+package roomescape.global.config.interceptor;
 
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.Cookie;
@@ -6,8 +6,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.auth.JwtTokenProvider;
-import roomescape.auth.exception.NotFoundCookieException;
+import roomescape.global.auth.JwtTokenProvider;
+import roomescape.global.auth.exception.NotFoundCookieException;
 import roomescape.global.exception.ForbiddenException;
 import roomescape.user.domain.Role;
