+++ src/main/java/roomescape/global/config/resolver/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.config.resolver;
+package roomescape.global.config.resolver;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,8 +8,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.auth.JwtTokenProvider;
-import roomescape.auth.service.AuthService;
+import roomescape.global.auth.JwtTokenProvider;
+import roomescape.global.auth.service.AuthService;
 import roomescape.user.domain.User;
 
 @Component
