+++ src/test/java/roomescape/admin/controller/AdminControllerTest.java
@@ -17,9 +17,9 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.admin.domain.dto.AdminReservationRequestDto;
 import roomescape.admin.fixture.AdminTestDataConfig;
-import roomescape.auth.domain.dto.TokenResponseDto;
+import roomescape.global.auth.domain.dto.TokenResponseDto;
 import roomescape.auth.fixture.AuthFixture;
-import roomescape.auth.service.AuthService;
+import roomescape.global.auth.service.AuthService;
 import roomescape.reservationtime.ReservationTimeTestDataConfig;
 import roomescape.theme.ThemeTestDataConfig;
 import roomescape.user.MemberTestDataConfig;
