+++ src/test/java/roomescape/auth/controller/AuthControllerTest.java
@@ -14,11 +14,11 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.auth.JwtTokenProvider;
-import roomescape.auth.domain.dto.TokenRequestDto;
-import roomescape.auth.domain.dto.TokenResponseDto;
+import roomescape.global.auth.JwtTokenProvider;
+import roomescape.global.auth.domain.dto.TokenRequestDto;
+import roomescape.global.auth.domain.dto.TokenResponseDto;
 import roomescape.auth.fixture.AuthFixture;
-import roomescape.auth.service.AuthService;
+import roomescape.global.auth.service.AuthService;
 import roomescape.user.domain.Role;
 import roomescape.user.domain.User;
 import roomescape.user.fixture.UserFixture;
