+++ src/test/java/roomescape/auth/service/AuthServiceTest.java
@@ -8,10 +8,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.auth.JwtTokenProvider;
-import roomescape.auth.domain.dto.TokenResponseDto;
-import roomescape.auth.exception.InvalidTokenException;
+import roomescape.global.auth.JwtTokenProvider;
+import roomescape.global.auth.domain.dto.TokenResponseDto;
+import roomescape.global.auth.exception.InvalidTokenException;
 import roomescape.auth.fixture.AuthFixture;
+import roomescape.global.auth.service.AuthService;
 import roomescape.user.domain.Role;
 import roomescape.user.domain.User;
 import roomescape.user.domain.dto.UserResponseDto;
