+++ src/test/java/roomescape/config/WebMvcConfigurationTest.java
@@ -10,10 +10,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import roomescape.admin.fixture.AdminTestDataConfig;
-import roomescape.auth.domain.dto.TokenRequestDto;
-import roomescape.auth.domain.dto.TokenResponseDto;
+import roomescape.global.auth.domain.dto.TokenRequestDto;
+import roomescape.global.auth.domain.dto.TokenResponseDto;
 import roomescape.auth.fixture.AuthFixture;
-import roomescape.auth.service.AuthService;
+import roomescape.global.auth.service.AuthService;
 import roomescape.user.MemberTestDataConfig;
 import roomescape.user.domain.User;
 import roomescape.user.fixture.AbstractUserTestDataConfig;
