+++ src/test/java/roomescape/reservation/repository/JdbcReservationRepositoryTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.common.KeyHolderManager;
+import roomescape.global.common.KeyHolderManager;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.fixture.ReservationFixture;
 import roomescape.reservationtime.domain.ReservationTime;
